+++ x/twap/export_test.go
@@ -15,6 +15,8 @@ type (
 	GeometricTwapStrategy  = geometric
 )
 
+var GeometricTwapMathBase = geometricTwapMathBase
+
 func (k Keeper) StoreNewRecord(ctx sdk.Context, record types.TwapRecord) {
 	k.storeNewRecord(ctx, record)
 }
